Added sprite scale support for AABB physics bodies.
This support assumes that the body indeed has a sprite reference...

Included a reference to the plugin with the facade. This makes it easier to remove the plugin at runtime; plugin.remove(this.game.slopes.plugin).

Simplified creating the slopes body because Phaser.Utils.mixin is a pain (hates null values).

Removed tile-specific solver support because it isn't used. SAT is, and will likely remain, the only solver.

Ensured that tilemap debug rendering only wraps if the layer has its wrap property set to true.

Updated typescript definitions.

Minor improvements to some documentation spacing in the Facade class.